update names  to refer to xtalk not cross_talk
