+++ src/pygama/evt/modules/geds.py
@@ -8,86 +8,11 @@ from collections.abc import Sequence
 from lgdo import types
 
 from .. import utils
-from . import cross_talk
+from . import xtalk
 from legendmeta import LegendMetadata
 
 
 
-
-
-def manipulate_ctx_matrix(cross_talk_matrix:dict,positive_cross_talk_matrix:dict=None,det_names:bool=False):
-    """
-    Function to read in and manipulate the cross talk matrix.
-    Parameters
-    ----------
-    xtalk_matrix_filename (str)
-         Path to the xtalk matrix
-    positive_xtalk_matrix_filename (str)
-        Path to the positive polarity cross talk matrix
-    det_names
-        bool to convert det names to rawids
-     
-    """
-  
-    
-    if (det_names is True):
-        cross_talk_matrix=convert_matrix_det_names_to_rawid(cross_talk_matrix)
-
-    # read the positive matrix
-    if (positive_cross_talk_matrix is not None):
-        
-        if (det_names is True):
-            positive_cross_talk_matrix=convert_matrix_det_names_to_rawid(positive_cross_talk_matrix)
-        
-        # merge +ive and -ive matrix
-        matrix_merge={}
-        for key_row,row in cross_talk_matrix.items():
-            matrix_merge[key_row]={}
-            for key_col,data in row.items():
-                
-                positive_ctx = positive_cross_talk_matrix[key_row][key_col]
-                negative_ctx = data
-
-                if (positive_ctx>negative_ctx):
-                    matrix_merge[key_row][key_col]=-positive_ctx
-                else:
-                    matrix_merge[key_row][key_col]=negative_ctx
-
-        cross_talk_matrix=matrix_merge
-
-    return cross_talk_matrix
-
-def convert_matrix_det_names_to_rawid(matrix:dict)->dict:
-    """
-    Converts a cross talk matrix with keys of detector names to one with keys rawids
-    Parameters
-    ----------
-    matrix
-        dictonary of the cross talk matrix
-    """
-    metadb = LegendMetadata()
-    chmap = metadb.channelmap("20230323T000000Z")
-
-    geds_mapping = {
-         _name :f"ch{_dict['daq']['rawid']}"
-        for _name, _dict in chmap.items()
-        if chmap[_name]["system"] == "geds"
-    }
-    matrix_conv={}
-    for key_row,row in matrix.items():
-
-        if key_row not in geds_mapping.keys():
-            raise ValueError(f"channel {key_row} doesnt have a valid rawid")
-
-        matrix_conv[geds_mapping[key_row]]={}
-        for key_col,data in row.items():
-
-            if key_col not in geds_mapping.keys():
-                raise ValueError(f"channel {key_col} doesnt have a valid rawid")
-
-            matrix_conv[geds_mapping[key_row]][geds_mapping[key_col]]=data
-    
-    return matrix_conv
 
 
 def apply_xtalk_correction(
@@ -139,6 +64,7 @@ def apply_xtalk_correction(
     except FileNotFoundError:
          raise ValueError(f"path to x-talk matrix {xtalk_matrix_filename} does not exist")
     
+    
     positive_xtalk_matrix=None
     if (positive_xtalk_matrix_filename is not None):
         try:
@@ -147,13 +73,13 @@ def apply_xtalk_correction(
         except FileNotFoundError:
             raise ValueError(f"path to x-talk matrix {positive_xtalk_matrix_filename} does not exist")
 
-    cross_talk_matrix=manipulate_ctx_matrix(xtalk_matrix,positive_xtalk_matrix,det_names)
+    xtalk_matrix=xtalk.manipulate_xtalk_matrix(xtalk_matrix,positive_xtalk_matrix,det_names)
 
     # do the correction
-    energies_corr = cross_talk.cross_talk_corrected_energy_awkard_slow(
+    energies_corr = xtalk.xtalk_corrected_energy_awkard_slow(
         energies=energy_observable.view_as("ak"),
         rawids=rawids.view_as("ak"),
-        matrix=cross_talk_matrix,
+        matrix=xtalk_matrix,
         allow_non_existing=False,
         threshold=threshold,
     )
